Adicionando TTY gem para loading no console

+++ modulo_03/crypto_wallet/lib/tasks/dev.rake
@@ -1,7 +1,14 @@
 namespace :dev do
   desc "Configura o ambiente de desenvolvimento do projeto"
   task setup: :environment do
-    %x(rails db:drop:_unsafe db:create db:migrate db:seed)
+    if Rails.env.development?
+      spinner = TTY::Spinner.new("[:spinner] Executando tarefas...")
+      spinner.auto_spin
+      %x(rails db:drop:_unsafe db:create db:migrate db:seed)
+      spinner.success("[Concluido]")
+    else
+      puts "Função habilitada somente para ambiente de Desenvolvimento"
+    end
   end
 
 end
